First push

+++ CrocCleaner/CrocCleanerDlg.cpp
@@ -1,58 +1,18 @@
-
-// CrocCleanerDlg.cpp : implementation file
-//
-
 #include "pch.h"
-#include "framework.h"
 #include "CrocCleaner.h"
 #include "CrocCleanerDlg.h"
-#include "afxdialogex.h"
-#include <thread>
+
+
 
 #ifdef _DEBUG
 #define new DEBUG_NEW
 #endif
 
 
-// CAboutDlg dialog used for App About
-
-class CAboutDlg : public CDialogEx
-{
-public:
-	CAboutDlg();
-
-// Dialog Data
-#ifdef AFX_DESIGN_TIME
-	enum { IDD = IDD_ABOUTBOX };
-#endif
-
-protected:
-	virtual void DoDataExchange(CDataExchange* pDX);    // DDX/DDV support
-
-// Implementation
-protected:
-	DECLARE_MESSAGE_MAP()
-};
-
-CAboutDlg::CAboutDlg() : CDialogEx(IDD_ABOUTBOX)
-{
-}
-
-void CAboutDlg::DoDataExchange(CDataExchange* pDX)
-{
-	CDialogEx::DoDataExchange(pDX);
-}
-
-BEGIN_MESSAGE_MAP(CAboutDlg, CDialogEx)
-END_MESSAGE_MAP()
-
-
-// CCrocCleanerDlg dialog
-
-
-
 CCrocCleanerDlg::CCrocCleanerDlg(CWnd* pParent /*=nullptr*/)
-	: CDialogEx(IDD_CROCCLEANER_DIALOG, pParent),iFileCounter{0},iFileNumber{0}
+	: CDialogEx(IDD_CROCCLEANER_DIALOG, pParent),
+	iFileCounter(0),
+	iFileNumber(0)
 {
 	m_hIcon = AfxGetApp()->LoadIcon(IDR_MAINFRAME);
 }
@@ -78,15 +38,10 @@ BEGIN_MESSAGE_MAP(CCrocCleanerDlg, CDialogEx)
 END_MESSAGE_MAP()
 
 
-// CCrocCleanerDlg message handlers
-
 BOOL CCrocCleanerDlg::OnInitDialog()
 {
 	CDialogEx::OnInitDialog();
 
-	// Add "About..." menu item to system menu.
-
-	// IDM_ABOUTBOX must be in the system command range.
 	ASSERT((IDM_ABOUTBOX & 0xFFF0) == IDM_ABOUTBOX);
 	ASSERT(IDM_ABOUTBOX < 0xF000);
 
@@ -104,21 +59,17 @@ BOOL CCrocCleanerDlg::OnInitDialog()
 		}
 	}
 
-	// Set the icon for this dialog.  The framework does this automatically
-	//  when the application's main window is not a dialog
-	SetIcon(m_hIcon, TRUE);			// Set big icon
-	SetIcon(m_hIcon, FALSE);		// Set small icon
+	SetIcon(m_hIcon, TRUE);		
+	SetIcon(m_hIcon, FALSE);	
 
-	// TODO: Add extra initialization here
 	m_btnFilesList.ShowWindow(false);
 	m_btnFilesList.EnableWindow(false);
 	m_brwsPath.EnableFolderBrowseButton();
-	cpBar.SetRange(0, 100);
-	iFileCounter = 0;
-	iFileNumber = 0;
-	
 
-	return TRUE;  // return TRUE  unless you set the focus to a control
+	cpBar.SetPos(0);
+	cpBar.SetRange(0, 100);
+
+	return TRUE;  
 }
 
 void CCrocCleanerDlg::OnSysCommand(UINT nID, LPARAM lParam)
@@ -134,19 +85,14 @@ void CCrocCleanerDlg::OnSysCommand(UINT nID, LPARAM lParam)
 	}
 }
 
-// If you add a minimize button to your dialog, you will need the code below
-//  to draw the icon.  For MFC applications using the document/view model,
-//  this is automatically done for you by the framework.
-
 void CCrocCleanerDlg::OnPaint()
 {
 	if (IsIconic())
 	{
-		CPaintDC dc(this); // device context for painting
+		CPaintDC dc(this); 
 
 		SendMessage(WM_ICONERASEBKGND, reinterpret_cast<WPARAM>(dc.GetSafeHdc()), 0);
 
-		// Center icon in client rectangle
 		int cxIcon = GetSystemMetrics(SM_CXICON);
 		int cyIcon = GetSystemMetrics(SM_CYICON);
 		CRect rect;
@@ -154,7 +100,6 @@ void CCrocCleanerDlg::OnPaint()
 		int x = (rect.Width() - cxIcon + 1) / 2;
 		int y = (rect.Height() - cyIcon + 1) / 2;
 
-		// Draw the icon
 		dc.DrawIcon(x, y, m_hIcon);
 	}
 	else
@@ -163,63 +108,30 @@ void CCrocCleanerDlg::OnPaint()
 	}
 }
 
-// The system calls this function to obtain the cursor to display while the user drags
-//  the minimized window.
 HCURSOR CCrocCleanerDlg::OnQueryDragIcon()
 {
 	return static_cast<HCURSOR>(m_hIcon);
 }
 
-class ThreadData {
-public:
-	int iFileCounterPtr;
-	int iFileNumberPtr;
-	CProgressCtrl* progressBarPointer;
-	ThreadData::ThreadData(int cnt, int nmb, CProgressCtrl* pbp) 
-		:iFileCounterPtr{ cnt }, iFileNumberPtr{ nmb }, progressBarPointer{ pbp }{}
-};
-
-UINT CCrocCleanerDlg::threadControlFunction(LPVOID pParam) {
-
-	CCrocCleanerDlg* mainWindowPtr = (CCrocCleanerDlg*)pParam;
-
-	mainWindowPtr->cpBar.SetRange(0, mainWindowPtr->iFileNumber);
-	
-	while (mainWindowPtr->iFileCounter <= mainWindowPtr->iFileNumber)
-	{
-		mainWindowPtr->cpBar.SetPos(mainWindowPtr->iFileCounter);
-		::UpdateWindow(mainWindowPtr->GetSafeHwnd());
-	};
-
-	
-	return 0;
-}
-
 void CCrocCleanerDlg::OnBnClickedbtnclean()
 {
-	
-	//Getting the path and creating and object 
 	CString path;
 	m_brwsPath.GetWindowTextW(path);
-	std::string stringPath(CW2A(path.GetString()));
-	FD.setPath(stringPath);
-	//Initialise variables
-	cpBar.SetPos(0);
-	cpBar.SetRange(0, 100);
-	this->iFileNumber = 0;
-	this->iFileCounter = 0;
 
-	//Getting the selection
+	std::string stringPath(CW2A(path.GetString()));
+
+	FD.setPath(stringPath);
+
+	cpBar.SetPos(0);
+
 	int userSelection = cComboBoxSelection.GetCurSel();
+
 	if (!path.IsEmpty())
 	{
-		if (userSelection >= 0) //0
+		if (userSelection >= 0)
 		{
+			auto iClearCatalogues = static_cast<bool>(cCatalogueClear.GetState());
 
-			//Getting the aditional catalogue deleting 
-			int iClearCatalogues = cCatalogueClear.GetState();
-
-			//Getting the time span
 			CTime timeStartDate{ 0 };
 			CTime timeEndDate{ 0 };
 			CString scStartDate, scEndDate;
@@ -228,48 +140,49 @@ void CCrocCleanerDlg::OnBnClickedbtnclean()
 			dwResult = cDtmStartDate.GetTime(timeStartDate);
 			dwResult = cDtmEndDate.GetTime(timeEndDate);
 
-			scStartDate = timeStartDate.Format(_T("%Y-%m-%d")); //"%Y-%m-%d %H:%M:%S"
-			scEndDate = timeEndDate.Format(_T("%Y-%m-%d"));	//"%Y-%m-%d %H:%M:%S"
+			scStartDate = timeStartDate.Format(_T("%Y-%m-%d")); 
+			scEndDate = timeEndDate.Format(_T("%Y-%m-%d"));	
+
 			std::string sStartDate(CW2A(scStartDate.GetString()));
 			std::string sEndDate(CW2A(scEndDate.GetString()));
 
-			//Find files
 			files = FD.findRequiredFiles(userSelection, sStartDate, sEndDate);
-			this->iFileNumber = files.size();
-			this->iFileCounter = 0;
+			
+			if (files.size() == 0)
+			{
+				AfxMessageBox(L"No files found.", MB_ICONINFORMATION);
+				return;
+			}
 
-			//Prepare progess bar thread and its data
-			::AfxBeginThread(threadControlFunction, this);
+			iFileNumber = files.size();
+			iFileCounter = 0;
 
-			//Delete all found files
-			FD.deletePaths(files, &iFileCounter);
+			cpBar.SetRange(0,static_cast<short>(iFileNumber));
 
-			//Deleting empty catalogues if needed
+			FD.deletePaths(files, cpBar);
+
 			if (iClearCatalogues == 1) {
 				FD.deleteEmptyCatalogues();
 			}
 			
 			m_btnFilesList.ShowWindow(true);
 			m_btnFilesList.EnableWindow(true);
-			::AfxMessageBox(L"Files deleted", MB_ICONINFORMATION);
-
+			AfxMessageBox(L"Files deleted", MB_ICONINFORMATION);
 		}
 		else
 		{
 			m_btnFilesList.ShowWindow(false);
 			m_btnFilesList.EnableWindow(false);
-			::AfxMessageBox(L"Wrong searching type selection");
+			AfxMessageBox(L"Wrong searching type selection");
 		}
 	}
 	else
 	{
 		m_btnFilesList.ShowWindow(false);
 		m_btnFilesList.EnableWindow(false);
-		::AfxMessageBox(L"Path is not selected");
+		AfxMessageBox(L"Path is not selected");
 	}
 }
-
-
 
 void CCrocCleanerDlg::OnBnClickedbtnfileslist()
 {
@@ -278,3 +191,15 @@ void CCrocCleanerDlg::OnBnClickedbtnfileslist()
 	FLDlg.printVector();
 	FLDlg.DoModal();
 }
+
+CAboutDlg::CAboutDlg() : CDialogEx(IDD_ABOUTBOX)
+{
+}
+
+void CAboutDlg::DoDataExchange(CDataExchange* pDX)
+{
+	CDialogEx::DoDataExchange(pDX);
+}
+
+BEGIN_MESSAGE_MAP(CAboutDlg, CDialogEx)
+END_MESSAGE_MAP()
